backend/requirements.txt
# Azure Functions
azure-functions>=1.17.0

# FastAPI
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Supabase
supabase>=2.0.0

# Database
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
pgvector>=0.2.4

# LlamaIndex
llama-index>=0.9.0
llama-index-embeddings-openai>=0.1.0
llama-index-llms-openai>=0.1.0

# OpenAI
openai>=1.6.0

# PDF処理
pypdf>=3.17.0
pdfplumber>=0.10.0

# ユーティリティ
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.0
tenacity>=8.2.0

# 開発用
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
